requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
findspark==2.0.1
idna==3.10
kafka-python==2.2.6
py4j==0.10.9.7
pyspark==3.5.5
requests==2.32.3
urllib3==2.4.0
